fix: 全面修复 130 个 Shader 的兼容性问题
- UNITY_MATRIX_MVP → UnityObjectToClipPos (实际代码替换)
- _Object2World → unity_ObjectToWorld
- _World2Object → unity_WorldToObject
- 添加 #pragma target 3.0/4.0
- fixed → half 精度替换
- 删除 BindChannels 块
- Surface Shader Lambert/BlinnPhong → Standard
- 删除未使用的 _MainTex 属性
- 共修复 124 个文件，216 处问题

+++ Assets/Effect/Dissolve2/Resources/dissolve.shader
@@ -118,7 +118,7 @@
 		half _Glossiness;
 		half _Metallic;
 		float _DissolveColorWidth, _Brightness, _Cutoff, _Smoothness;
-		fixed4 _Color, _DisColor;
+		half4 _Color, _DisColor;
 
 		// Add instancing support for this shader. You need to check 'Enable Instancing' on materials that use the shader.
 		// See https://docs.unity3d.com/Manual/GPUInstancing.html for more information about instancing.
@@ -132,7 +132,7 @@
 		{
 			_DisAmount *= _Shrink;
 			// Main texture tinted by color
-			fixed4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
+			half4 c = tex2D(_MainTex, IN.uv_MainTex) * _Color;
 
 			float3 blendNormal = saturate(pow(IN.worldNormal * 1.4, 4));
