+++ Assets/GeometryShader/Example/Res/Shader/jianciShader.shader
@@ -45,7 +45,7 @@
 			sampler2D _MainTex;
 			float4 _MainTex_ST;
 			float _Length;
-			fixed4 _CustomColor;
+			half4 _CustomColor;
       
 			//添加顶点
 			void ADD_VERT(float3 v, g2f o, inout TriangleStream<g2f> tristream)
@@ -108,9 +108,9 @@
 				ADD_TRI(point3,point2,cebterPoint3,o,tristream);
 			}
  
-			fixed4 frag (g2f i) : SV_Target
+			half4 frag (g2f i) : SV_Target
 			{
-				fixed4 col = tex2D(_MainTex, i.uv) * _CustomColor;
+				half4 col = tex2D(_MainTex, i.uv) * _CustomColor;
 				return col;
 			}
